Consolidate napari-mcp-bridge into main package
- Moved all napari-mcp-bridge components into src/napari_mcp/
- Created unified package structure with base.py, server.py, bridge_server.py, and widget.py
- Updated pyproject.toml to include napari plugin entry points
- Fixed all test imports to use new napari_mcp package structure
- Removed duplicate napari-mcp-bridge directory
- Maintained all functionality as both CLI tool and napari plugin
