--- napari-mcp-bridge/pyproject.toml
@@ -1,46 +1,0 @@
-[project]
-name = "napari-mcp-bridge"
-version = "0.1.0"
-description = "Napari plugin to expose current viewer as MCP server"
-readme = "README.md"
-requires-python = ">=3.10"
-license = {text = "MIT"}
-authors = [
-    {name = "Ilan Theodoro", email = "ilan.silva@czbiohub.org"},
-]
-keywords = ["napari", "plugin", "mcp", "ai", "claude"]
-classifiers = [
-    "Development Status :: 3 - Alpha",
-    "Framework :: napari",
-    "Intended Audience :: Science/Research",
-    "License :: OSI Approved :: MIT License",
-    "Operating System :: OS Independent",
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.12",
-    "Programming Language :: Python :: 3.13",
-    "Topic :: Scientific/Engineering :: Image Processing",
-]
-dependencies = [
-    "napari>=0.5.5",
-    "fastmcp>=2.7.0",
-    "qtpy>=2.4.1",
-    "pyqt6>=6.5.0",
-    "numpy>=1.26.0",
-    "Pillow>=10.3.0",
-    "imageio>=2.34.0",
-]
-
-[project.entry-points."napari.manifest"]
-napari-mcp-bridge = "napari_mcp_bridge:napari.yaml"
-
-[build-system]
-requires = ["setuptools>=61.0", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[tool.setuptools.packages.find]
-where = ["src"]
-
-[tool.setuptools.package-data]
-napari_mcp_bridge = ["napari.yaml"]
