+++ pyproject.toml
@@ -14,6 +14,7 @@ maintainers = [
 keywords = ["napari", "mcp", "microscopy", "image-analysis", "ai", "claude"]
 classifiers = [
     "Development Status :: 3 - Alpha",
+    "Framework :: napari",
     "Intended Audience :: Science/Research",
     "Intended Audience :: Developers",
     "License :: OSI Approved :: MIT License",
@@ -64,17 +65,20 @@ all = [
 ]
 
 [project.scripts]
-napari-mcp = "napari_mcp_server:main"
+napari-mcp = "napari_mcp.server:main"
+
+[project.entry-points."napari.manifest"]
+napari-mcp = "napari_mcp:napari.yaml"
 
 [build-system]
 requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
-[tool.setuptools]
-py-modules = ["napari_mcp_server"]
+[tool.setuptools.packages.find]
+where = ["src"]
 
-[tool.setuptools.package-dir]
-"" = "src"
+[tool.setuptools.package-data]
+napari_mcp = ["napari.yaml"]
 
 [tool.pytest.ini_options]
 addopts = "-q -p no:napari"
