+++ src/napari_mcp/base.py
