+++ src/napari_mcp/bridge_server.py
