+++ src/napari_mcp/server.py
